+++ src/app/(frontend)/layout.tsx
@@ -5,14 +5,13 @@ import { GeistMono } from 'geist/font/mono'
 import { GeistSans } from 'geist/font/sans'
 import React from 'react'
 
-import { AdminBar } from '@/components/AdminBar'
 import { Footer } from '@/Footer/Component'
 import { Header } from '@/Header/Component'
 import { Providers } from '@/providers'
 import { InitTheme } from '@/providers/Theme/InitTheme'
 import { mergeOpenGraph } from '@/utilities/mergeOpenGraph'
-import { draftMode } from 'next/headers'
 import { StarfieldClient } from '@/components/StarfieldClient'
+import { PayloadAdminRemover } from '@/components/PayloadAdminRemover'
 
 import './globals.css'
 import { getServerSideURL } from '@/utilities/getURL'
@@ -25,8 +24,6 @@ const fontOptimizationStyles = `
 `
 
 export default async function RootLayout({ children }: { children: React.ReactNode }) {
-  const { isEnabled } = await draftMode()
-
   return (
     <html className={cn(GeistSans.variable, GeistMono.variable)} lang="en" suppressHydrationWarning>
       <head>
@@ -46,12 +43,7 @@ export default async function RootLayout({ children }: { children: React.ReactNo
         <StarfieldClient />
 
         <Providers>
-          <AdminBar
-            adminBarProps={{
-              preview: isEnabled,
-            }}
-          />
-
+          <PayloadAdminRemover />
           <Header />
           {children}
           <Footer />